+++ app/javascript/application.js
@@ -1,4 +1,4 @@
 // Entry point for the build script in your package.json
 import "@hotwired/turbo-rails";
-import "./controllers/index.js";
+import "./controllers";
 import "bootstrap";
